extract alias utils from "helpers"
also removed utils tests, since the use cases are covered via alias tests
