+++ addon-test-support/-private/helpers.js
@@ -1,4 +1,3 @@
-import { get } from '@ember/object';
 import Ceibo from 'ceibo';
 import deprecate from './deprecate';
 export { default as $ } from 'jquery';
@@ -214,62 +213,6 @@ export function findClosestValue(node, property) {
   if (isPresent(parent)) {
     return findClosestValue(parent, property);
   }
-}
-
-/**
- * @public
- *
- * Returns a boolean indicating whether an object contains a given property.
- * The path to a nested property should be indicated by a dot-separated string.
- *
- * @param {Object} object - object to check for the target property
- * @param {string} pathToProp - dot-separated path to property
- * @return {Boolean}
- */
-export function objectHasProperty(object, pathToProp) {
-  const pathSegments = pathToProp.split('.');
-
-  for (let i = 0; i < pathSegments.length; i++) {
-    const key = pathSegments[i];
-    if (object === null || object === undefined || !object.hasOwnProperty(key)) {
-      return false;
-    } else {
-      object = object[key];
-    }
-  }
-
-  return true;
-}
-
-/**
- * @public
- *
- * Returns the value of an object property. If the property is a function,
- * the return value is that function bound to its "owner."
- *
- * @param {Object} object - object on which to look up the target property
- * @param {string} pathToProp - dot-separated path to property
- * @return {Boolean|String|Number|Function|Null|Undefined} - value of property
- */
-export function getProperty(object, pathToProp) {
-  const pathSegments = pathToProp.split('.');
-
-  if (pathSegments.length === 1) {
-    const value = get(object, pathToProp);
-    return typeof value === 'function' ? value.bind(object) : value;
-  }
-
-  const pathToPropOwner = pathSegments.slice(0, -1).join('.');
-  const propOwner = get(object, pathToPropOwner);
-
-  if (propOwner === null || propOwner === undefined) {
-    return undefined;
-  }
-
-  const propKey = pathSegments[pathSegments.length - 1];
-  const value = get(propOwner, propKey);
-
-  return typeof value === 'function' ? value.bind(propOwner) : value;
 }
 
 export function isPageObject(property) {
