+++ addon-test-support/macros/alias.js
@@ -1,9 +1,6 @@
+import { get } from '@ember/object';
 import { throwBetterError } from '../-private/better-errors';
-import {
-  getProperty,
-  objectHasProperty
-} from '../-private/helpers';
-import { chainable } from "../-private/chainable";
+import { chainable } from '../-private/chainable';
 
 const ALIASED_PROP_NOT_FOUND = 'PageObject does not contain aliased property';
 
@@ -79,7 +76,7 @@ const ALIASED_PROP_NOT_FOUND = 'PageObject does not contain aliased property';
  */
 export function alias(pathToProp, options = {}) {
   return {
-    isDescriptor: true,
+  isDescriptor: true,
 
     get(key) {
       if (!objectHasProperty(this, pathToProp)) {
@@ -102,4 +99,60 @@ export function alias(pathToProp, options = {}) {
       };
     }
   };
+}
+
+/**
+ * @private
+ *
+ * Returns a boolean indicating whether an object contains a given property.
+ * The path to a nested property should be indicated by a dot-separated string.
+ *
+ * @param {Object} object - object to check for the target property
+ * @param {string} pathToProp - dot-separated path to property
+ * @return {Boolean}
+ */
+function objectHasProperty(object, pathToProp) {
+  const pathSegments = pathToProp.split('.');
+
+  for (let i = 0; i < pathSegments.length; i++) {
+    const key = pathSegments[i];
+    if (object === null || object === undefined || !object.hasOwnProperty(key)) {
+      return false;
+    } else {
+      object = object[key];
+    }
+  }
+
+  return true;
+}
+
+/**
+ * @private
+ *
+ * Returns the value of an object property. If the property is a function,
+ * the return value is that function bound to its "owner."
+ *
+ * @param {Object} object - object on which to look up the target property
+ * @param {string} pathToProp - dot-separated path to property
+ * @return {Boolean|String|Number|Function|Null|Undefined} - value of property
+ */
+function getProperty(object, pathToProp) {
+  const pathSegments = pathToProp.split('.');
+
+  if (pathSegments.length === 1) {
+    const value = get(object, pathToProp);
+    return typeof value === 'function' ? value.bind(object) : value;
+  }
+
+  const pathToPropOwner = pathSegments.slice(0, -1).join('.');
+  const propOwner = get(object, pathToPropOwner);
+
+  if (propOwner === null || propOwner === undefined) {
+    return undefined;
+  }
+
+  const propKey = pathSegments[pathSegments.length - 1];
+  const value = get(propOwner, propKey);
+
+  return typeof value === 'function' ? value.bind(propOwner) : value;
 }
